Added depth, lookout confusing
Added depth to each player function that looksout for special. This is to prevent the  recursive loop from going on forever. 
Confused at the lookout mechanic and where they shouldbe looking if the busdriver switches the target. They look at the  house  i.e their point er is point  to the same memory address but the memory address now  holds a different value.