redid Detective, added blockers  and helper function
Redid detective to account for the structure that's newly created. The lookout doesn't seem to be right at the moment but will have to double check with mods. Added blocking roles. Added a helper function to check whether target is blocked. Tested on the Shooter role but not any other roles.