Fixed lookout, blockers*, blocked ability, sort function added
The preaction function didnt pass information back when blocked so action like mass murder would carry on.  converted to afunction and conditional statement is checked. 
Lookout is broken because of how the busdriver switches the person. The fix is counterintuitive. The lookout iull not switch to lookout the new target. Rather the lookout sees who visits the original person as anyone who looks at the original person looks at the switched person. 
Blockers described  as A>B>C>D should be a stack data structure. This is not how mafia is programmed. 
so it is by  the order described 
The function sort by rand and sort by role enable a randomization though this is not good because  clear board is now hard coded to the results column (12). 
This should be moved to the first columns, Would require alot of push over though. Shold be done